requirements.txt
google-genai
selenium
undetected-chromedriver
python-dotenv
requests
loguru
